adding readme and examples
