--- SampleApp/.cr/personal/FavoritesList/List.xml
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Root Type="DevExpress.CodeRush.Foundation.CodePlaces.Options.FavoritesListContainer">
-  <Options Language="Neutral">
-    <Groups />
-  </Options>
-</Root>
