--- SampleApp/README.md
@@ -1,9 +1,0 @@
-# Sample App for Pre-recorded Transcription
-
-To run the sample app, follow a few simple steps.
-
-1. Provide your API key in `Program.cs`: `var credentials = new Credentials("");`
-1. From the `SampleApp/` directory, run the command: `dotnet run SampleApp.csproj`
-1. The sample file is transcribed and the JSON response is printed to the terminal.
-
-Use this sample app to learn the basics of making Deepgram API calls via the .NET SDK! Reuse and adapt the code for your own projects.
