+++ examples/README.md
@@ -1,0 +1,25 @@
+# Examples for Testing Features Locally
+
+Test features locally or use these examples as quickstarts to get up and running with the Deeegram .NET SDK.
+
+Here are the steps to follow to run the examples:
+
+## Add Your Code (if you are testing changes to the SDK)
+
+Make your changes to the SDK (be sure you are on a branch you have created to do this work).
+
+### Edit the API key
+
+Inside the example file, replace the API key where it says 'YOUR_DEEPGRAM_API_KEY'
+
+`DEEPGRAM_API_KEY = 'YOUR_DEEPGRAM_API_KEY'`
+
+### Run the project
+
+Make sure you're in the directory with the `Program.cs` file and run the project with the following command.
+
+`dotnet run Program.cs`
+
+### After testing
+
+After you have used the example files to test your code, be sure to reset the example file to the way it was when you started (i.e. discard features you may have added to the options dictionary when testing features).
